backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0
pydantic>=2.8.0
pydantic-settings>=2.0.0
google-generativeai>=0.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.0.0
pypdf>=3.17.0
